rename folder provider to zabbix
