+++ zabbix/helper.go
