+++ zabbix/resource_zabbix_trigger_test.go
